Only owner can delete/edit images - resolved

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -5,6 +5,7 @@ import ImageHoster.model.Tag;
 import ImageHoster.model.User;
 import ImageHoster.service.ImageService;
 import ImageHoster.service.TagService;
+import ImageHoster.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -26,6 +27,11 @@ public class ImageController {
 
     @Autowired
     private TagService tagService;
+
+    //Annotated to Userservice.
+    @Autowired
+    private UserService userService;
+
 
     //This method displays all the images in the user home page after successful login
     @RequestMapping("images")
@@ -92,13 +98,20 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model, HttpSession session) {
         Image image = imageService.getImage(imageId);
-
-        String tags = convertTagsToString(image.getTags());
         model.addAttribute("image", image);
+        if(userService.isUserLoggedInUser(image.getUser(),session)){
+            //Added all tags
+            String tags = convertTagsToString(image.getTags());
+            model.addAttribute("tags", tags);
+            return "images/edit";
+        }
+        List<Tag> tags = image.getTags();
         model.addAttribute("tags", tags);
-        return "images/edit";
+        String error = "Only the owner of the image can edit/delete the image";
+        model.addAttribute("editError", error);
+        return "images/image";
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -140,9 +153,21 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId,HttpSession session, Model model) {
+        Image image = imageService.getImage(imageId);
+
+        // if user is Userlogged in then delete
+        if(userService.isUserLoggedInUser(image.getUser(),session)){
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        }
+        String error = "Only the owner of the image can edit/delete the image";
+        model.addAttribute("deleteError", error);
+        model.addAttribute("image", image);
+        //Added tags
+        List<Tag> tags = image.getTags();
+        model.addAttribute("tags", tags);
+        return "images/image";
     }
